code cleanup.
