+++ ServiceApp/src/org/droidplanner/services/android/core/drone/autopilot/CommonMavLinkDrone.java
@@ -3,8 +3,10 @@ package org.droidplanner.services.android.core.drone.autopilot;
 import android.text.TextUtils;
 
 import com.MAVLink.Messages.MAVLinkMessage;
+import com.MAVLink.common.msg_attitude;
 import com.MAVLink.common.msg_radio_status;
 import com.o3dr.services.android.lib.drone.attribute.AttributeType;
+import com.o3dr.services.android.lib.drone.property.Attitude;
 import com.o3dr.services.android.lib.drone.property.Battery;
 import com.o3dr.services.android.lib.drone.property.DroneAttribute;
 import com.o3dr.services.android.lib.drone.property.Signal;
@@ -14,6 +16,7 @@ import com.o3dr.services.android.lib.util.MathUtils;
 import org.droidplanner.services.android.core.MAVLink.MAVLinkStreams;
 import org.droidplanner.services.android.core.drone.DroneEvents;
 import org.droidplanner.services.android.core.drone.DroneInterfaces;
+import org.droidplanner.services.android.utils.CommonApiUtils;
 
 /**
  * Base drone implementation. Supports mavlink messages belonging to the common set: https://pixhawk.ethz.ch/mavlink/
@@ -27,6 +30,7 @@ public abstract class CommonMavLinkDrone implements MavLinkDrone {
     protected final Speed speed = new Speed();
     protected final Battery battery = new Battery();
     protected final Signal signal = new Signal();
+    protected final Attitude attitude = new Attitude();
 
     protected CommonMavLinkDrone(DroneInterfaces.Handler handler, MAVLinkStreams.MAVLinkOutputStream mavClient) {
         this.MavClient = mavClient;
@@ -78,6 +82,9 @@ public abstract class CommonMavLinkDrone implements MavLinkDrone {
 
             case AttributeType.SIGNAL:
                 return signal;
+
+            case AttributeType.ATTITUDE:
+                return attitude;
         }
 
         return null;
@@ -91,7 +98,25 @@ public abstract class CommonMavLinkDrone implements MavLinkDrone {
                 processSignalUpdate(m_radio_status.rxerrors, m_radio_status.fixed, m_radio_status.rssi,
                         m_radio_status.remrssi, m_radio_status.txbuf, m_radio_status.noise, m_radio_status.remnoise);
                 break;
+
+            case msg_attitude.MAVLINK_MSG_ID_ATTITUDE:
+                msg_attitude m_att = (msg_attitude) message;
+                processAttitude(m_att);
+                break;
         }
+    }
+
+    private void processAttitude(msg_attitude m_att) {
+        attitude.setRoll(CommonApiUtils.fromRadToDeg(m_att.roll));
+        attitude.setRollSpeed(CommonApiUtils.fromRadToDeg(m_att.rollspeed));
+
+        attitude.setPitch(CommonApiUtils.fromRadToDeg(m_att.pitch));
+        attitude.setPitchSpeed(CommonApiUtils.fromRadToDeg(m_att.pitchspeed));
+
+        attitude.setYaw(CommonApiUtils.fromRadToDeg(m_att.yaw));
+        attitude.setYawSpeed(CommonApiUtils.fromRadToDeg(m_att.yawspeed));
+
+        notifyDroneEvent(DroneInterfaces.DroneEventsType.ATTITUDE);
     }
 
     protected void processSignalUpdate(int rxerrors, int fixed, short rssi, short remrssi, short txbuf,
